# docs/swagger.yaml
basePath: /api/v1
definitions:
  sdk.Candle:
    properties:
      close:
        description: Close price of candlestick
        example: 214.21
        type: number
      high:
        description: High price of candlestick
        example: 214.21
        type: number
      low:
        description: Low price of candlestick
        example: 214.21
        type: number
      ns:
        description: Nanoseconds portion of the event timestamp
        example: 123456
        type: integer
      open:
        description: Open price of candlestick
        example: 214.21
        type: number
      pub:
        description: DataBento Publisher ID
        example: 1
        type: integer
      sym:
        description: Ticker of the trade
        example: AAPL
        type: string
      ts:
        description: Trade event timestamp as seconds from the epoch
        example: 1713644400
        type: integer
      volume:
        description: Volume in candlestick
        example: 100
        type: integer
    type: object
  sdk.TradeTick:
    properties:
      mkt:
        description: Market of the trade
        example: Q
        type: string
      ns:
        description: Nanoseconds portion of the event timestamp
        example: 123456
        type: integer
      pub:
        description: DataBento Publisher ID
        example: 1
        type: integer
      px:
        description: Trade price
        example: 214.21
        type: number
      sym:
        description: Ticker of the trade
        example: AAPL
        type: string
      sz:
        description: Trade size/volume
        example: 100
        type: integer
      ts:
        description: Trade event timestamp as seconds from the epoch
        example: 1713644400
        type: integer
    type: object
host: api.example.com
info:
  contact:
    email: nosupport@nimble.markets
    name: Neomantra Corp
    url: https://nimble.markets
  description: DuckDB-backed DBN Golang web service
  license:
    name: MIT
    url: https://mit-license.org
  termsOfService: file:///dev/null
  title: dbn-duckduck-goose
  version: "1.0"
paths:
  /candles/{dataset}/{ticker}:
    get:
      description: Returns a time range of OHLCV for a Dataset and Ticker
      operationId: GetOhlcvByDatasetAndTicker
      parameters:
      - description: DataBento dataset
        example: DBEQ.BASIC
        in: path
        name: dataset
        required: true
        type: string
      - description: symbol to query
        example: AAPL
        in: path
        name: ticker
        required: true
        type: string
      - description: (optional) start of date range in ISO8601. Default is midnight
          Eastern.
        format: ISO8601
        in: query
        name: start
        type: string
      - description: (optional) end of date range in ISO8601. Default is now.
        format: ISO8601
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: array of Candles
          schema:
            items:
              $ref: '#/definitions/sdk.Candle'
            type: array
        "404":
          description: dataset not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get a time range of OHLCV for a Dataset and Ticker
  /charts/candles/{dataset}/{ticker}:
    get:
      description: Returns an HTML page candlestick chart with volume and EMA for
        the given dataset and ticker.
      operationId: GetCandleChartByDatasetAndTicker
      parameters:
      - description: DataBento dataset
        example: DBEQ.BASIC
        in: path
        name: dataset
        required: true
        type: string
      - description: symbol to query
        example: AAPL
        in: path
        name: ticker
        required: true
        type: string
      - description: (optional) start of date range in ISO8601. Default is midnight
          Eastern.
        format: ISO8601
        in: query
        name: start
        type: string
      - description: (optional) end of date range in ISO8601. Default is now.
        format: ISO8601
        in: query
        name: end
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page with candlestick chart
          schema:
            type: string
        "404":
          description: dataset not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Returns an HTML page candlestick chart with volume and EMA for the
        given dataset and ticker.
  /last-trades/csv/{dataset}/{ticker}:
    get:
      description: Returns Excel file with last N trades by dataset and ticker.
      operationId: GetLastTradesByDatasetAndTickerCSV
      parameters:
      - description: DataBento dataset
        example: DBEQ.BASIC
        in: path
        name: dataset
        required: true
        type: string
      - description: symbol to query
        example: AAPL
        in: path
        name: ticker
        required: true
        type: string
      - description: (optional) number of trades to return - default is 25
        in: query
        name: count
        type: integer
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file with last N trades
          schema:
            type: string
        "404":
          description: dataset not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GET last N trades by market and ticker
  /last-trades/excel/{dataset}/{ticker}:
    get:
      description: Returns Excel file with last N trades by dataset and ticker.
      operationId: GetLastTradesByDatasetAndTickerExcel
      parameters:
      - description: DataBento dataset
        example: DBEQ.BASIC
        in: path
        name: dataset
        required: true
        type: string
      - description: symbol to query
        example: AAPL
        in: path
        name: ticker
        required: true
        type: string
      - description: (optional) number of trades to return - default is 25
        in: query
        name: count
        type: integer
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file with last N trades
          schema:
            type: string
        "404":
          description: dataset not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GET last N trades by market and ticker
  /last-trades/json/{dataset}/{ticker}:
    get:
      description: Returns the last N trades by dataset and ticker.
      operationId: GetLastTradesByDatasetAndTicker
      parameters:
      - description: DataBento dataset
        example: DBEQ.BASIC
        in: path
        name: dataset
        required: true
        type: string
      - description: symbol to query
        example: AAPL
        in: path
        name: ticker
        required: true
        type: string
      - description: (optional) number of trades to return - default is 25
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: array of TradeTicks
          schema:
            items:
              $ref: '#/definitions/sdk.TradeTick'
            type: array
        "404":
          description: dataset not found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: GET last N trades by market and ticker
schemes:
- http
swagger: "2.0"